Internal and cmd prompt restructure
