--- internal/commands/readme.md
@@ -1,1 +1,0 @@
-# `./commands`
